move logic from utils.go
