+++ destination.go
@@ -18,10 +18,14 @@ package pinecone
 
 import (
 	"context"
+	"encoding/json"
 	"fmt"
+	"net/url"
 
 	sdk "github.com/conduitio/conduit-connector-sdk"
 	"github.com/pinecone-io/go-pinecone/pinecone"
+	"github.com/pkg/errors"
+	"google.golang.org/protobuf/types/known/structpb"
 )
 
 type Destination struct {
@@ -101,4 +105,118 @@ func (d *Destination) Teardown(ctx context.Context) error {
 		return fmt.Errorf("failed to close index: %w", err)
 	}
 	return nil
+}
+
+func newIndex(ctx context.Context, config DestinationConfig) (*pinecone.IndexConnection, error) {
+	client, err := pinecone.NewClient(pinecone.NewClientParams{
+		ApiKey: config.APIKey,
+	})
+	if err != nil {
+		return nil, fmt.Errorf("error creating Pinecone client: %w", err)
+	}
+	sdk.Logger(ctx).Info().Msg("created pinecone client")
+
+	hostURL, err := url.Parse(config.Host)
+	if err != nil {
+		return nil, fmt.Errorf("invalid host url: %w", err)
+	}
+
+	var index *pinecone.IndexConnection
+	if config.Namespace != "" {
+		index, err = client.IndexWithNamespace(hostURL.Host, config.Namespace)
+		if err != nil {
+			return nil, fmt.Errorf(
+				"error establishing index connection to namespace %v: %w",
+				config.Namespace, err)
+		}
+	} else {
+		index, err = client.Index(hostURL.Host)
+		if err != nil {
+			return nil, fmt.Errorf("error establishing index connection: %w", err)
+		}
+	}
+	sdk.Logger(ctx).Info().Msg("created pinecone index")
+
+	return index, nil
+}
+
+func recordID(key sdk.Data) string {
+	return string(key.Bytes())
+}
+
+type sparseValues struct {
+	Indices []uint32  `json:"indices"`
+	Values  []float32 `json:"values"`
+}
+
+type pineconeVectorValues struct {
+	Values       []float32    `json:"values"`
+	SparseValues sparseValues `json:"sparse_values,omitempty"`
+}
+
+func (r pineconeVectorValues) PineconeSparseValues() *pinecone.SparseValues {
+	// the used pinecone go client needs a nil pointer when no sparse values given, or else it
+	// will throw a "Sparse vector must contain at least one value" error
+	if len(r.SparseValues.Indices) == 0 && len(r.SparseValues.Values) == 0 {
+		return nil
+	}
+
+	v := &pinecone.SparseValues{
+		Indices: r.SparseValues.Indices,
+		Values:  r.SparseValues.Values,
+	}
+	return v
+}
+
+func parsePineconeVector(rec sdk.Record) (*pinecone.Vector, error) {
+	id := recordID(rec.Key)
+
+	payload, err := parsePineconeVectorValues(rec)
+	if err != nil {
+		return nil, err
+	}
+
+	metadata, err := parsePineconeMetadata(rec)
+	if err != nil {
+		return nil, err
+	}
+
+	vec := &pinecone.Vector{
+		//revive:disable-next-line
+		Id:           id,
+		Values:       payload.Values,
+		SparseValues: payload.PineconeSparseValues(),
+		Metadata:     metadata,
+	}
+
+	return vec, nil
+}
+
+func parsePineconeVectorValues(rec sdk.Record) (values pineconeVectorValues, err error) {
+	data := rec.Payload.After
+
+	if data == nil || len(data.Bytes()) == 0 {
+		return values, errors.New("empty payload")
+	}
+
+	err = json.Unmarshal(data.Bytes(), &values)
+	if err != nil {
+		return values, fmt.Errorf("error unmarshalling JSON: %w", err)
+	}
+
+	return values, nil
+}
+
+func parsePineconeMetadata(rec sdk.Record) (*pinecone.Metadata, error) {
+	convertedMap := make(map[string]any)
+	for key, value := range rec.Metadata {
+		convertedMap[key] = value
+	}
+
+	metadata, err := structpb.NewStruct(convertedMap)
+	if err != nil {
+		return nil, fmt.Errorf("error creating metadata: %w", err)
+	}
+
+	return metadata, nil
 }
